project/settings.py
import os

from dotenv import load_dotenv, find_dotenv

load_dotenv(find_dotenv())

DATABASES = {
    'default': {
        'ENGINE': 'django.db.backends.postgresql_psycopg2',
        'HOST': os.getenv('CHECKPOINT_DB_HOST'),
        'PORT': os.getenv('CHECKPOINT_DB_PORT'),
        'NAME': os.getenv('CHECKPOINT_DB_NAME'),
        'USER': os.getenv('CHECKPOINT_DB_USER'),
        'PASSWORD': os.getenv('CHECKPOINT_DB_PASSWORD'),
    }
}

INSTALLED_APPS = ['datacenter']

SECRET_KEY = os.getenv('DJANGO_SECRET_KEY')

DEBUG = True

ROOT_URLCONF = "project.urls"

ALLOWED_HOSTS = ['*']


BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
TEMPLATES = [
    {
        'BACKEND': 'django.template.backends.django.DjangoTemplates',
        'DIRS': [os.path.join(BASE_DIR, 'templates')],
        'APP_DIRS': True,
    },
]


USE_L10N = True

LANGUAGE_CODE = 'ru-ru'

TIME_ZONE = 'Europe/Moscow'

USE_TZ = True
